Organize code into core and gui packages
